@@ -108,9 +108,12 @@ def chat_response_to_completion_response(
     chat_response: ChatResponse,
 ) -> CompletionResponse:
     """Convert a chat response to a completion response."""
+    additional_kwargs = chat_response.message.additional_kwargs
+    additional_kwargs.update(chat_response.additional_kwargs)
+
     return CompletionResponse(
         text=chat_response.message.content or "",
-        additional_kwargs=chat_response.message.additional_kwargs,
+        additional_kwargs=additional_kwargs,
         raw=chat_response.raw,
     )
 
@@ -122,9 +125,12 @@ def stream_chat_response_to_completion_response(
 
     def gen() -> CompletionResponseGen:
         for response in chat_response_gen:
+            additional_kwargs = response.message.additional_kwargs
+            additional_kwargs.update(response.additional_kwargs)
+
             yield CompletionResponse(
                 text=response.message.content or "",
-                additional_kwargs=response.message.additional_kwargs,
+                additional_kwargs=additional_kwargs,
                 delta=response.delta,
                 raw=response.raw,
             )
@@ -284,9 +290,12 @@ def astream_chat_response_to_completion_response(
 
     async def gen() -> CompletionResponseAsyncGen:
         async for response in chat_response_gen:
+            additional_kwargs = response.message.additional_kwargs
+            additional_kwargs.update(response.additional_kwargs)
+
             yield CompletionResponse(
                 text=response.message.content or "",
-                additional_kwargs=response.message.additional_kwargs,
+                additional_kwargs=additional_kwargs,
                 delta=response.delta,
                 raw=response.raw,
             )
