@@ -2,6 +2,7 @@ import logging
 from typing import Any, Callable, List
 
 from llama_index.core.node_parser.interface import TextSplitter
+from llama_index.core.utils import globals_helper
 
 logger = logging.getLogger(__name__)
 
@@ -79,10 +80,9 @@ def split_by_sentence_tokenizer_internal(text: str, tokenizer: Any) -> List[str]
 
 
 def split_by_sentence_tokenizer() -> Callable[[str], List[str]]:
-    import nltk
-
-    tokenizer = nltk.tokenize.PunktSentenceTokenizer()
-    return lambda text: split_by_sentence_tokenizer_internal(text, tokenizer)
+    return lambda text: split_by_sentence_tokenizer_internal(
+        text, globals_helper.punkt_tokenizer
+    )
 
 
 def split_by_regex(regex: str) -> Callable[[str], List[str]]:
