@@ -371,22 +371,21 @@ class MemgraphPropertyGraphStore(PropertyGraphStore):
         params = [r.dict() for r in relations]
         for index in range(0, len(params), CHUNK_SIZE):
             chunked_params = params[index : index + CHUNK_SIZE]
-            for param in chunked_params:
-                formatted_properties = ", ".join(
-                    [f"{key}: {value!r}" for key, value in param["properties"].items()]
-                )
-                self.structured_query(
-                    f"""
-                    MERGE (source: {BASE_NODE_LABEL} {{id: "{param["source_id"]}"}})
-                    ON CREATE SET source:Chunk
-                    MERGE (target: {BASE_NODE_LABEL} {{id: "{param["target_id"]}"}})
-                    ON CREATE SET target:Chunk
-                    WITH source, target
-                    MERGE (source)-[r:{param["label"]}]->(target)
-                    SET r += {{{formatted_properties}}}
-                    RETURN count(*)
-                    """
-                )
+
+            self.structured_query(
+                f"""
+                UNWIND $data AS row
+                MERGE (source: {BASE_NODE_LABEL} {{id: row.source_id}})
+                ON CREATE SET source:Chunk
+                MERGE (target: {BASE_NODE_LABEL} {{id: row.target_id}})
+                ON CREATE SET target:Chunk
+                WITH source, target, row
+                CREATE (source)-[r:row.label]->(target)
+                SET r += row.properties
+                RETURN count(*)
+                """,
+                param_map={"data": chunked_params},
+            )
 
     def get(
         self,
@@ -1053,10 +1052,19 @@ class MemgraphPropertyGraphStore(PropertyGraphStore):
             # Check if vector index is configured
             try:
                 self.structured_query(
-                    """CALL vector_search.show_index_info() YIELD * RETURN *;"""
+                    f"""
+                    CREATE VECTOR INDEX {VECTOR_INDEX_NAME} ON :{BASE_ENTITY_LABEL}(embedding) WITH CONFIG {{"dimension": 1536, "capacity": 1000}};
+                    """
                 )
                 self._supports_vector_index = True
-                return
+                logger.info(
+                    "Vector index %s was created with a fixed embedding dimension of 1536. "
+                    "If your chosen LLM model uses a different dimension, manually create the vector index with the following query:\n"
+                    'CREATE VECTOR INDEX %s ON :%s(embedding) WITH CONFIG {"dimension": <INSERT_DIMENSION>, "capacity": 1000};',
+                    VECTOR_INDEX_NAME,
+                    VECTOR_INDEX_NAME,
+                    BASE_ENTITY_LABEL,
+                )
             except neo4j.exceptions.Neo4jError as decode_error:
                 self._supports_vector_index = False
                 if (
@@ -1065,10 +1073,8 @@ class MemgraphPropertyGraphStore(PropertyGraphStore):
                     and "vector_search.show_index_info" in decode_error.message
                 ):
                     logger.info(
-                        """To use vector indices and vector search, start
-                        Memgraph with the experimental vector search feature
-                        flag and configure vector index. Falling back to
-                        alternative queries."""
+                        """Failed to create vector index entity:
+                        Given vector index already exists."""
                     )
         else:
             self._supports_vector_index = False
