@@ -5,11 +5,13 @@ import typing
 from typing import (
     TYPE_CHECKING,
     Any,
+    AsyncGenerator,
     Dict,
     Generator,
     List,
     Optional,
     Sequence,
+    Type,
     Union,
 )
 
@@ -36,9 +38,9 @@ from llama_index.core.callbacks import CallbackManager
 from llama_index.core.constants import DEFAULT_TEMPERATURE, DEFAULT_NUM_OUTPUTS
 from llama_index.core.llms.callbacks import llm_chat_callback, llm_completion_callback
 from llama_index.core.llms.function_calling import FunctionCallingLLM
-from llama_index.core.llms.llm import ToolSelection
+from llama_index.core.llms.llm import ToolSelection, Model
 from llama_index.core.prompts import PromptTemplate
-from llama_index.core.types import Model
+from llama_index.core.program.utils import FlexibleModel
 from llama_index.llms.google_genai.utils import (
     chat_from_gemini_response,
     chat_message_to_gemini,
@@ -425,21 +427,20 @@ class GoogleGenAI(FunctionCallingLLM):
     @dispatcher.span
     def structured_predict_without_function_calling(
         self,
-        output_cls: type[BaseModel],
+        output_cls: Type[Model],
         prompt: PromptTemplate,
         llm_kwargs: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> BaseModel:
+        **prompt_args: Any,
+    ) -> Model:
         """Structured predict."""
         llm_kwargs = llm_kwargs or {}
-        all_kwargs = {**llm_kwargs, **kwargs}
 
-        messages = prompt.format_messages()
+        messages = prompt.format_messages(**prompt_args)
         response = self._client.models.generate_content(
             model=self.model,
             contents=list(map(chat_message_to_gemini, messages)),
             **{
-                **all_kwargs,
+                **llm_kwargs,
                 **{
                     "config": {
                         "response_mime_type": "application/json",
@@ -457,89 +458,85 @@ class GoogleGenAI(FunctionCallingLLM):
     @dispatcher.span
     def structured_predict(
         self,
-        output_cls: type[BaseModel],
+        output_cls: Type[Model],
         prompt: PromptTemplate,
         llm_kwargs: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> BaseModel:
+        **prompt_args: Any,
+    ) -> Model:
         """Structured predict."""
         llm_kwargs = llm_kwargs or {}
-        all_kwargs = {**llm_kwargs, **kwargs}
 
         if self.is_function_calling_model:
             llm_kwargs["tool_choice"] = (
                 "required"
-                if "tool_choice" not in all_kwargs
-                else all_kwargs["tool_choice"]
+                if "tool_choice" not in llm_kwargs
+                else llm_kwargs["tool_choice"]
             )
 
         return super().structured_predict(
-            output_cls, prompt, llm_kwargs=llm_kwargs, **kwargs
+            output_cls, prompt, llm_kwargs=llm_kwargs, **prompt_args
         )
 
     @dispatcher.span
     async def astructured_predict(
         self,
-        output_cls: type[BaseModel],
+        output_cls: Type[Model],
         prompt: PromptTemplate,
         llm_kwargs: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> BaseModel:
+        **prompt_args: Any,
+    ) -> Model:
         """Structured predict."""
         llm_kwargs = llm_kwargs or {}
-        all_kwargs = {**llm_kwargs, **kwargs}
 
         if self.is_function_calling_model:
             llm_kwargs["tool_choice"] = (
                 "required"
-                if "tool_choice" not in all_kwargs
-                else all_kwargs["tool_choice"]
+                if "tool_choice" not in llm_kwargs
+                else llm_kwargs["tool_choice"]
             )
 
         return await super().astructured_predict(
-            output_cls, prompt, llm_kwargs=llm_kwargs, **kwargs
+            output_cls, prompt, llm_kwargs=llm_kwargs, **prompt_args
         )
 
     @dispatcher.span
     def stream_structured_predict(
         self,
-        output_cls: type[BaseModel],
+        output_cls: Type[Model],
         prompt: PromptTemplate,
         llm_kwargs: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> Generator[Union[Model, List[Model]], None, None]:
+        **prompt_args: Any,
+    ) -> Generator[Union[Model, FlexibleModel], None, None]:
         """Stream structured predict."""
         llm_kwargs = llm_kwargs or {}
-        all_kwargs = {**llm_kwargs, **kwargs}
 
         if self.is_function_calling_model:
             llm_kwargs["tool_choice"] = (
                 "required"
-                if "tool_choice" not in all_kwargs
-                else all_kwargs["tool_choice"]
+                if "tool_choice" not in llm_kwargs
+                else llm_kwargs["tool_choice"]
             )
         return super().stream_structured_predict(
-            output_cls, prompt, llm_kwargs=llm_kwargs, **kwargs
+            output_cls, prompt, llm_kwargs=llm_kwargs, **prompt_args
         )
 
     @dispatcher.span
     async def astream_structured_predict(
         self,
-        output_cls: type[BaseModel],
+        output_cls: Type[Model],
         prompt: PromptTemplate,
         llm_kwargs: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ) -> Generator[Union[Model, List[Model]], None, None]:
+        **prompt_args: Any,
+    ) -> AsyncGenerator[Union[Model, FlexibleModel], None]:
         """Stream structured predict."""
         llm_kwargs = llm_kwargs or {}
-        all_kwargs = {**llm_kwargs, **kwargs}
 
         if self.is_function_calling_model:
             llm_kwargs["tool_choice"] = (
                 "required"
-                if "tool_choice" not in all_kwargs
-                else all_kwargs["tool_choice"]
+                if "tool_choice" not in llm_kwargs
+                else llm_kwargs["tool_choice"]
             )
         return await super().astream_structured_predict(
-            output_cls, prompt, llm_kwargs=llm_kwargs, **kwargs
+            output_cls, prompt, llm_kwargs=llm_kwargs, **prompt_args
         )
