@@ -123,7 +123,7 @@ class IpexLLM(CustomLLM):
     is_chat_model: bool = Field(
         default=False,
         description=(
-            LLMMetadata.__fields__["is_chat_model"].field_info.description
+            LLMMetadata.__fields__["is_chat_model"].description
             + " Be sure to verify that you either pass an appropriate tokenizer "
             "that can convert prompts to properly formatted chat messages or a "
             "`messages_to_prompt` that does so."
