@@ -1,5 +1,5 @@
 import logging
-from typing import Any, Callable, Dict, List, Optional, Sequence, Type
+from typing import Any, Callable, Dict, List, Optional, Sequence, Type, Union
 
 from llama_index.core.base.llms.types import ChatMessage
 from llama_index.core.bridge.pydantic import BaseModel
@@ -13,7 +13,16 @@ from tenacity import (
 )
 
 import litellm
+from litellm import completion
 from litellm.utils import Message
+from litellm.main import ModelResponse, CustomStreamWrapper
+from litellm.types.utils import ChatCompletionDeltaToolCall
+from llama_index.core.base.llms.types import (
+    TextBlock,
+    ImageBlock,
+    AudioBlock,
+)
+
 
 MISSING_API_KEY_ERROR_MESSAGE = """No API key found for LLM.
 E.g. to use openai Please set the OPENAI_API_KEY environment variable or \
@@ -49,13 +58,13 @@ def _create_retry_decorator(max_retries: int) -> Callable[[Any], Any]:
 
 
 def completion_with_retry(is_chat_model: bool, max_retries: int, **kwargs: Any) -> Any:
-    from litellm import completion
-
     """Use tenacity to retry the completion call."""
     retry_decorator = _create_retry_decorator(max_retries=max_retries)
 
     @retry_decorator
-    def _completion_with_retry(**kwargs: Any) -> Any:
+    def _completion_with_retry(
+        **kwargs: Any,
+    ) -> Union[ModelResponse, CustomStreamWrapper]:
         return completion(**kwargs)
 
     return _completion_with_retry(**kwargs)
@@ -70,7 +79,9 @@ async def acompletion_with_retry(
     retry_decorator = _create_retry_decorator(max_retries=max_retries)
 
     @retry_decorator
-    async def _completion_with_retry(**kwargs: Any) -> Any:
+    async def _completion_with_retry(
+        **kwargs: Any,
+    ) -> Union[ModelResponse, CustomStreamWrapper]:
         # Use OpenAI's async api https://github.com/openai/openai-python#async-api
         return await acompletion(**kwargs)
 
@@ -125,10 +136,14 @@ def is_chat_model(model: str) -> bool:
     return model in litellm.model_list
 
 
-def is_function_calling_model(model: str) -> bool:
-    is_chat_model_ = is_chat_model(model)
-    is_old = "0314" in model or "0301" in model
-    return is_chat_model_ and not is_old
+def is_function_calling_model(
+    model: str, custom_llm_provider: Optional[str] = None
+) -> bool:
+    import litellm
+
+    return litellm.supports_function_calling(
+        model, custom_llm_provider=custom_llm_provider
+    )
 
 
 def get_completion_endpoint(is_chat_model: bool) -> CompletionClientType:
@@ -137,16 +152,63 @@ def get_completion_endpoint(is_chat_model: bool) -> CompletionClientType:
     return completion
 
 
-def to_openai_message_dict(message: ChatMessage) -> dict:
-    """Convert generic message to OpenAI message dict."""
+def to_openailike_message_dict(message: ChatMessage) -> dict:
+    """Convert a ChatMessage to an OpenAI-like message dict."""
+    content = []
+    content_txt = ""
+    for block in message.blocks:
+        if isinstance(block, TextBlock):
+            content.append({"type": "text", "text": block.text})
+            content_txt += block.text
+        elif isinstance(block, ImageBlock):
+            if block.url:
+                content.append(
+                    {
+                        "type": "image_url",
+                        "image_url": {
+                            "url": str(block.url),
+                            "detail": block.detail or "auto",
+                        },
+                    }
+                )
+            else:
+                img_bytes = block.resolve_image(as_base64=True).read()
+                img_str = img_bytes.decode("utf-8")
+                content.append(
+                    {
+                        "type": "image_url",
+                        "image_url": {
+                            "url": f"data:{block.image_mimetype};base64,{img_str}",
+                            "detail": block.detail or "auto",
+                        },
+                    }
+                )
+        elif isinstance(block, AudioBlock):
+            audio_bytes = block.resolve_audio(as_base64=True).read()
+            audio_str = audio_bytes.decode("utf-8")
+            content.append(
+                {
+                    "type": "input_audio",
+                    "input_audio": {
+                        "data": audio_str,
+                        "format": block.format,
+                    },
+                }
+            )
+        else:
+            msg = f"Unsupported content block type: {type(block).__name__}"
+            raise ValueError(msg)
+
+    # use content as text instead of blocks if all blocks are text. Safer for backwards compatibility with older openai-like APIs
     message_dict = {
-        "role": message.role,
-        "content": message.content,
+        "role": message.role.value,
+        "content": (
+            content_txt
+            if all(isinstance(block, TextBlock) for block in message.blocks)
+            else content
+        ),
     }
 
-    # NOTE: openai messages have additional arguments:
-    # - function messages have `name`
-    # - assistant messages have optional `function_call`
     message_dict.update(message.additional_kwargs)
 
     return message_dict
@@ -154,7 +216,7 @@ def to_openai_message_dict(message: ChatMessage) -> dict:
 
 def to_openai_message_dicts(messages: Sequence[ChatMessage]) -> List[dict]:
     """Convert generic messages to OpenAI message dicts."""
-    return [to_openai_message_dict(message) for message in messages]
+    return [to_openailike_message_dict(message) for message in messages]
 
 
 def from_openai_message_dict(message_dict: dict) -> ChatMessage:
@@ -175,8 +237,12 @@ def from_litellm_message(message: Message) -> ChatMessage:
     role = message.get("role")
     # NOTE: Azure OpenAI returns function calling messages without a content key
     content = message.get("content", None)
+    tool_calls = message.get("tool_calls")
+    additional_kwargs = {}
+    if tool_calls:
+        additional_kwargs["tool_calls"] = tool_calls
 
-    return ChatMessage(role=role, content=content)
+    return ChatMessage(role=role, content=content, additional_kwargs=additional_kwargs)
 
 
 def from_openai_message_dicts(message_dicts: Sequence[dict]) -> List[ChatMessage]:
@@ -202,3 +268,109 @@ def validate_litellm_api_key(
     api_key = litellm.validate_environment()
     if api_key is None:
         raise ValueError(MISSING_API_KEY_ERROR_MESSAGE)
+
+
+def update_tool_calls(
+    tool_calls: List[dict],
+    tool_call_deltas: Optional[List[ChatCompletionDeltaToolCall]],
+) -> List[dict]:
+    """
+    Update the list of tool calls with deltas.
+
+    Args:
+        tool_calls: The current list of tool calls
+        tool_call_deltas: A list of deltas to update tool_calls with
+
+    Returns:
+        List[dict]: The updated tool calls
+    """
+    if not tool_call_deltas:
+        return tool_calls
+
+    for tool_call_delta in tool_call_deltas:
+        # Convert ChatCompletionDeltaToolCall to dict
+        delta_dict = {}
+        if tool_call_delta.id is not None:
+            delta_dict["id"] = tool_call_delta.id
+        if tool_call_delta.type is not None:
+            delta_dict["type"] = tool_call_delta.type
+        if hasattr(tool_call_delta, "index"):
+            delta_dict["index"] = tool_call_delta.index
+
+        if (
+            hasattr(tool_call_delta, "function")
+            and tool_call_delta.function is not None
+        ):
+            delta_dict["function"] = {}
+            if (
+                hasattr(tool_call_delta.function, "name")
+                and tool_call_delta.function.name is not None
+            ):
+                delta_dict["function"]["name"] = tool_call_delta.function.name
+            if (
+                hasattr(tool_call_delta.function, "arguments")
+                and tool_call_delta.function.arguments is not None
+            ):
+                delta_dict["function"]["arguments"] = tool_call_delta.function.arguments
+
+        if len(tool_calls) == 0:
+            # First tool call
+            tool_calls.append(delta_dict)
+        else:
+            # Try to find an existing tool call to update
+            found_match = False
+            for existing_tool in tool_calls:
+                # Check by index if available
+                index_match = False
+                if "index" in delta_dict and "index" in existing_tool:
+                    index_match = delta_dict["index"] == existing_tool["index"]
+
+                # Check by id if available
+                id_match = False
+                if "id" in delta_dict and "id" in existing_tool:
+                    id_match = delta_dict["id"] == existing_tool["id"]
+
+                if index_match or id_match:
+                    found_match = True
+                    # Update existing tool call
+                    if "function" in delta_dict:
+                        if "function" not in existing_tool:
+                            existing_tool["function"] = {}
+
+                        # Update function name if present
+                        if "name" in delta_dict["function"]:
+                            if "name" not in existing_tool["function"]:
+                                existing_tool["function"]["name"] = ""
+                            existing_tool["function"]["name"] += delta_dict[
+                                "function"
+                            ].get("name", "")
+
+                        # Update arguments if present
+                        if "arguments" in delta_dict["function"]:
+                            if "arguments" not in existing_tool["function"]:
+                                existing_tool["function"]["arguments"] = ""
+                            existing_tool["function"]["arguments"] += delta_dict[
+                                "function"
+                            ].get("arguments", "")
+
+                    # Update ID if present
+                    if "id" in delta_dict:
+                        if "id" not in existing_tool:
+                            existing_tool["id"] = ""
+                        existing_tool["id"] += delta_dict.get("id", "")
+
+                    # Update type if present
+                    if "type" in delta_dict:
+                        existing_tool["type"] = delta_dict["type"]
+
+                    # Update index if present
+                    if "index" in delta_dict:
+                        existing_tool["index"] = delta_dict["index"]
+
+                    break
+
+            # If no match was found, add as a new tool call
+            if not found_match and ("id" in delta_dict or "index" in delta_dict):
+                tool_calls.append(delta_dict)
+
+    return tool_calls
