@@ -1,3 +1,4 @@
 from llama_index.llms.openai.base import AsyncOpenAI, OpenAI, SyncOpenAI, Tokenizer
+from llama_index.llms.openai.responses import OpenAIResponses
 
-__all__ = ["OpenAI", "Tokenizer", "SyncOpenAI", "AsyncOpenAI"]
+__all__ = ["OpenAI", "OpenAIResponses", "Tokenizer", "SyncOpenAI", "AsyncOpenAI"]
