@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import List, Dict
+from typing import List, Dict, Any
+import uuid
 import sys
 import logging
 from enum import Enum
@@ -15,15 +16,31 @@ from typing import (
     Optional,
     Union,
 )
+
 from llama_index.core.vector_stores.types import (
     FilterCondition,
-)
-from llama_index.core.vector_stores.types import (
     MetadataFilters,
     FilterOperator,
 )
+from llama_index.core.vector_stores.utils import DEFAULT_TEXT_KEY
+
 
 logger = logging.getLogger(__name__)
+
+try:
+    from pymilvus import Function as BaseMilvusBuiltInFunction
+    from pymilvus import FunctionType
+except ImportError:
+    error_info = (
+        "Cannot import built-in function from PyMilvus. "
+        "To use Milvus built-in function, "
+        "you may need to upgrade pymilvus to the latest version:\n"
+        "pip install pymilvus -U\n"
+    )
+    logger.warning(error_info)
+
+
+DEFAULT_SPARSE_EMBEDDING_KEY = "sparse_embedding"
 
 
 class FilterOperatorFunction(str, Enum):
@@ -228,3 +245,39 @@ class BGEM3SparseEmbeddingFunction(BaseSparseEmbeddingFunction):
 
 def get_default_sparse_embedding_function() -> BGEM3SparseEmbeddingFunction:
     return BGEM3SparseEmbeddingFunction()
+
+
+class BM25BuiltInFunction(BaseMilvusBuiltInFunction):
+    """BM25 function built in Milvus."""
+
+    def __init__(
+        self,
+        input_field_names: Union[str, List[str]] = DEFAULT_TEXT_KEY,
+        output_field_names: Union[str, List[str]] = DEFAULT_SPARSE_EMBEDDING_KEY,
+        function_name: str = None,
+        function_description: str = "",
+        function_params: Optional[Dict] = None,
+        analyzer_params: Optional[Dict[Any, Any]] = None,
+        enable_match: bool = False,
+    ):
+        if not function_name:
+            function_name = f"bm25_function_{str(uuid.uuid4())[:8]}"
+        super().__init__(
+            function_name,
+            FunctionType.BM25,
+            input_field_names,
+            output_field_names,
+            function_description,
+            function_params,
+        )
+        self.analyzer_params = analyzer_params
+        self.enable_match = enable_match
+
+    def get_field_kwargs(self) -> dict:
+        field_schema_kwargs: Dict[Any, Any] = {
+            "enable_analyzer": True,
+            "enable_match": self.enable_match,
+        }
+        if self.analyzer_params is not None:
+            field_schema_kwargs["analyzer_params"] = self.analyzer_params
+        return field_schema_kwargs
