@@ -23,6 +23,8 @@ from llama_index.vector_stores.milvus.utils import (
 import pytest_asyncio
 
 TEST_URI = "./milvus_test.db"
+DIM = 64
+COLLECTION_NAME = "test_collection"
 
 
 class MockSparseEmbeddingFunction(BaseSparseEmbeddingFunction):
@@ -228,8 +230,8 @@ class TestMilvusAsync:
     def vector_store(self, event_loop) -> MilvusVectorStore:
         yield MilvusVectorStore(
             uri=TEST_URI,
-            dim=64,
-            collection_name="test_collection",
+            dim=DIM,
+            collection_name=COLLECTION_NAME,
             embedding_field="embedding",
             id_field="id",
             similarity_metric="COSINE",
@@ -241,24 +243,24 @@ class TestMilvusAsync:
         node1 = TextNode(
             id_="n1",
             text="n1_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
         )
         node2 = TextNode(
             id_="n2",
             text="n2_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
         )
         await vector_store.async_add([node1, node2])
         await vector_store.adelete(ref_doc_id="n2")
         query_res = await vector_store.aclient.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )
         assert query_res[0]["count(*)"] == 1
         await vector_store.adelete(ref_doc_id="n3")
         query_res = await vector_store.aclient.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )
         assert query_res[0]["count(*)"] == 0
 
@@ -268,42 +270,42 @@ class TestMilvusAsync:
         node1 = TextNode(
             id_="n1",
             text="n1_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
         )
         node2 = TextNode(
             id_="n2",
             text="n2_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
         )
         await vector_store.async_add([node1, node2])
         await vector_store.adelete_nodes(node_ids=["n1"])
         query_res = await vector_store.aclient.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )
         assert query_res[0]["count(*)"] == 1
         await vector_store.adelete_nodes(node_ids=["n2"])
         query_res = await vector_store.aclient.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )
         assert query_res[0]["count(*)"] == 0
 
     async def test_milvus_clear(self, vector_store: MilvusVectorStore, event_loop):
         await vector_store.aclear()
-        assert not vector_store.client.has_collection("test_collection")
+        assert not vector_store.client.has_collection(COLLECTION_NAME)
 
     async def test_get_nodes(self, vector_store: MilvusVectorStore, event_loop):
         node1 = TextNode(
             id_="n1",
             text="n1_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
         )
         node2 = TextNode(
             id_="n2",
             text="n2_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
         )
         await vector_store.async_add([node1, node2])
@@ -318,24 +320,24 @@ class TestMilvusAsync:
         node1 = TextNode(
             id_="n1",
             text="n1_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
         )
         node2 = TextNode(
             id_="n2",
             text="n2_text",
-            embedding=[-0.5] * 64,  # opposite direction of node1's embedding
+            embedding=[-0.5] * DIM,  # opposite direction of node1's embedding
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
         )
         await vector_store.async_add([node1, node2])
 
-        query = VectorStoreQuery(query_embedding=[0.5] * 64, similarity_top_k=1)
+        query = VectorStoreQuery(query_embedding=[0.5] * DIM, similarity_top_k=1)
         result = await vector_store.aquery(query=query)
         assert len(result.nodes) == 1
         assert result.nodes[0].id_ == "n1"
         assert result.nodes[0].text == "n1_text"
 
-        query = VectorStoreQuery(query_embedding=[-0.5] * 64, similarity_top_k=1)
+        query = VectorStoreQuery(query_embedding=[-0.5] * DIM, similarity_top_k=1)
         result = await vector_store.aquery(query=query)
         assert len(result.nodes) == 1
         assert result.nodes[0].id_ == "n2"
@@ -362,7 +364,7 @@ class TestMilvusAsync:
         )
         await vector_store.async_add([node1, node2, node3])
         query = VectorStoreQuery(
-            query_embedding=[0.5] * 64,
+            query_embedding=[0.5] * DIM,
             similarity_top_k=2,
             mode=VectorStoreQueryMode.MMR,
         )
@@ -373,11 +375,44 @@ class TestMilvusAsync:
         assert result.nodes[1].id_ == "n2"
         assert result.nodes[1].text == "n2_text"
 
+    async def test_query_sparse_mode(self, event_loop):
+        vector_store = MilvusVectorStore(
+            uri=TEST_URI,
+            collection_name=COLLECTION_NAME,
+            overwrite=True,
+            enable_dense=False,
+            enable_sparse=True,
+            sparse_embedding_function=MockSparseEmbeddingFunction(),
+            consistency_level="Strong",
+        )
+        node1 = TextNode(
+            id_="n1",
+            text="n1_text",
+            # embedding=[0.5] * 64,
+            relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
+        )
+        node2 = TextNode(
+            id_="n2",
+            text="n2_text",
+            # embedding=[-0.5] * 64,  # opposite direction of node1's embedding
+            relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
+        )
+        await vector_store.async_add([node1, node2])
+        query = VectorStoreQuery(
+            query_str="mock_str",
+            similarity_top_k=1,
+            mode=VectorStoreQueryMode.SPARSE,
+        )
+        result = await vector_store.aquery(query=query)
+        assert len(result.nodes) == 1
+        assert result.nodes[0].id_ == "n1"
+        assert result.nodes[0].text == "n1_text"
+
     async def test_query_hybrid_mode(self, event_loop):
         vector_store = MilvusVectorStore(
             uri=TEST_URI,
-            dim=64,
-            collection_name="test_collection",
+            dim=DIM,
+            collection_name=COLLECTION_NAME,
             overwrite=True,
             enable_sparse=True,
             hybrid_ranker="RRFRanker",
@@ -388,18 +423,18 @@ class TestMilvusAsync:
         node1 = TextNode(
             id_="n1",
             text="n1_text",
-            embedding=[0.5] * 64,
+            embedding=[0.5] * DIM,
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
         )
         node2 = TextNode(
             id_="n2",
             text="n2_text",
-            embedding=[-0.5] * 64,  # opposite direction of node1's embedding
+            embedding=[-0.5] * DIM,  # opposite direction of node1's embedding
             relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n3")},
         )
         await vector_store.async_add([node1, node2])
         query = VectorStoreQuery(
-            query_embedding=[0.5] * 64,
+            query_embedding=[0.5] * DIM,
             query_str="mock_str",
             similarity_top_k=1,
             mode=VectorStoreQueryMode.HYBRID,
@@ -425,7 +460,7 @@ class TestMilvusAsync:
             TextNode(
                 id_=f"n{i}",
                 text=f"text_{i}",
-                embedding=[0.5] * 64,
+                embedding=[0.5] * DIM,
             )
             for i in range(3)
         ]
@@ -488,13 +523,53 @@ class TestMilvusSync:
         return MilvusVectorStore(
             uri=TEST_URI,
             dim=64,
-            collection_name="test_collection",
+            collection_name=COLLECTION_NAME,
             embedding_field="embedding",
             id_field="id",
             similarity_metric="COSINE",
             consistency_level="Strong",
             overwrite=True,
         )
+
+    def test_milvus_index_management(self, vector_store: MilvusVectorStore):
+        # List all indexes
+        indexes = vector_store.client.list_indexes(COLLECTION_NAME)
+        # Drop existing indexes
+        for index_name in indexes:
+            vector_store.client.drop_index(
+                collection_name=COLLECTION_NAME, index_name=index_name
+            )
+
+        # check index_management="no_validation"
+        vector_store_1 = MilvusVectorStore(
+            uri=TEST_URI,
+            collection_name=COLLECTION_NAME,
+            index_management="no_validation",
+        )
+        indexes_1 = vector_store_1.client.list_indexes(COLLECTION_NAME)
+        assert len(indexes_1) == 0
+
+        # check index_management="create_if_not_exists"
+        vector_store_2 = MilvusVectorStore(
+            uri=TEST_URI,
+            collection_name=COLLECTION_NAME,
+            index_management="create_if_not_exists",
+        )
+        indexes_2 = vector_store_2.client.list_indexes(COLLECTION_NAME)
+        assert len(indexes_2) > 0
+
+    def test_milvus_add(self, vector_store: MilvusVectorStore):
+        node1 = TextNode(
+            id_="n1",
+            text="n1_text",
+            embedding=[0.5] * 64,
+            relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="n2")},
+        )
+        vector_store.add([node1])
+        row_count = vector_store.client.query(
+            COLLECTION_NAME, output_fields=["count(*)"]
+        )[0]["count(*)"]
+        assert row_count == 1
 
     def test_milvus_delete(self, vector_store: MilvusVectorStore):
         node1 = TextNode(
@@ -512,12 +587,12 @@ class TestMilvusSync:
         vector_store.add([node1, node2])
         vector_store.delete(ref_doc_id="n2")
         row_count = vector_store.client.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )[0]["count(*)"]
         assert row_count == 1
         vector_store.delete(ref_doc_id="n3")
         row_count = vector_store.client.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )[0]["count(*)"]
         assert row_count == 0
 
@@ -537,18 +612,18 @@ class TestMilvusSync:
         vector_store.add([node1, node2])
         vector_store.delete_nodes(node_ids=["n1"])
         row_count = vector_store.client.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )[0]["count(*)"]
         assert row_count == 1
         vector_store.delete_nodes(node_ids=["n2"])
         row_count = vector_store.client.query(
-            "test_collection", output_fields=["count(*)"]
+            COLLECTION_NAME, output_fields=["count(*)"]
         )[0]["count(*)"]
         assert row_count == 0
 
     def test_milvus_clear(self, vector_store: MilvusVectorStore):
         vector_store.clear()
-        assert not vector_store.client.has_collection("test_collection")
+        assert not vector_store.client.has_collection(COLLECTION_NAME)
 
     def test_get_nodes(self, vector_store: MilvusVectorStore):
         node1 = TextNode(
@@ -632,7 +707,7 @@ class TestMilvusSync:
         vector_store = MilvusVectorStore(
             uri=TEST_URI,
             dim=64,
-            collection_name="test_collection",
+            collection_name=COLLECTION_NAME,
             overwrite=True,
             enable_sparse=True,
             hybrid_ranker="RRFRanker",
